Move AtlanticElectricalHeaterWithAdjustableTemperatureSetpoint to seperate implementation
