+++ custom_components/tahoma/climate.py
@@ -3,6 +3,9 @@
 from homeassistant.components.climate import DOMAIN as CLIMATE
 
 from .climate_devices.atlantic_electrical_heater import AtlanticElectricalHeater
+from .climate_devices.atlantic_electrical_heater_with_adjustable_temperature_setpoint import (
+    AtlanticElectricalHeaterWithAdjustableTemperatureSetpoint,
+)
 from .climate_devices.atlantic_pass_apcdhw import AtlanticPassAPCDHW
 from .climate_devices.dimmer_exterior_heating import DimmerExteriorHeating
 from .climate_devices.somfy_thermostat import SomfyThermostat
@@ -11,7 +14,7 @@ from .const import DOMAIN
 
 TYPE = {
     "AtlanticElectricalHeater": AtlanticElectricalHeater,
-    "AtlanticElectricalHeaterWithAdjustableTemperatureSetpoint": AtlanticElectricalHeater,
+    "AtlanticElectricalHeaterWithAdjustableTemperatureSetpoint": AtlanticElectricalHeaterWithAdjustableTemperatureSetpoint,
     "AtlanticPassAPCDHW": AtlanticPassAPCDHW,
     "DimmerExteriorHeating": DimmerExteriorHeating,
     "SomfyThermostat": SomfyThermostat,
